[CLIENT] Add response info to ServerError exceptions
Currently if a ServerError or any of its descendants are raised, the response
body is included as part of the exception's message string. For example:

    > client.indices.delete(index: 'wat')
    Elasticsearch::Transport::Transport::Errors::NotFound: [404] {"error":
    {"root_cause":[{"type":"index_not_found_exception",
    "reason":"no such index","resource.type":"index_or_alias",
    "resource.id":"wat","index_uuid":"_na_","index":"wat"}],
    "type":"index_not_found_exception","reason":"no such index",
    "resource.type":"index_or_alias","resource.id":"wat",
    "index_uuid":"_na_","index":"wat"},"status":404}

In this case, a user may want to ignore the exception since the non-existence
of the index is exactly their intention. However, it's hard for them to
directly access the reason for the error, as the JSON response is just part of
an unformatted string.

To make the underlying response date easier to access, this change adds an
optional `response` parameter to the ServerError family of exceptions,
allowing third-party code to more fully inspect the issue, and respond
appropriately.